Add builder factory to modify existing instances agnostically
