+++ wdtk-datamodel/src/main/java/org/wikidata/wdtk/datamodel/helpers/EntityDocumentBuilder.java
@@ -4,7 +4,7 @@ package org.wikidata.wdtk.datamodel.helpers;
  * #%L
  * Wikidata Toolkit Data Model
  * %%
- * Copyright (C) 2014 - 2015 Wikidata Toolkit Developers
+ * Copyright (C) 2014 - 2018 Wikidata Toolkit Developers
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,42 +20,22 @@ package org.wikidata.wdtk.datamodel.helpers;
  * #L%
  */
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-
 import org.wikidata.wdtk.datamodel.interfaces.EntityDocument;
 import org.wikidata.wdtk.datamodel.interfaces.EntityIdValue;
-import org.wikidata.wdtk.datamodel.interfaces.MonolingualTextValue;
-import org.wikidata.wdtk.datamodel.interfaces.PropertyIdValue;
-import org.wikidata.wdtk.datamodel.interfaces.Statement;
-import org.wikidata.wdtk.datamodel.interfaces.StatementDocument;
-import org.wikidata.wdtk.datamodel.interfaces.StatementGroup;
-import org.wikidata.wdtk.datamodel.interfaces.TermedDocument;
 
-/**
- * Abstract base class for builders that construct {@link EntityDocument}
- * objects.
- *
- * @author Markus Kroetzsch
- *
- * @param <T>
- *            the type of the eventual concrete builder implementation
- * @param <O>
- *            the type of the object that is being built
- */
-public abstract class EntityDocumentBuilder<T extends EntityDocumentBuilder<T, O>, O extends StatementDocument & TermedDocument>
-		extends AbstractDataObjectBuilder<T, O> {
-
+public abstract class EntityDocumentBuilder<T extends EntityDocumentBuilder<T, O>, O extends EntityDocument>
+extends AbstractDataObjectBuilder<T, O>  {
+	
 	final EntityIdValue entityIdValue;
-	final ArrayList<MonolingualTextValue> labels = new ArrayList<>();
-	final ArrayList<MonolingualTextValue> descriptions = new ArrayList<>();
-	final ArrayList<MonolingualTextValue> aliases = new ArrayList<>();
-	final HashMap<PropertyIdValue, ArrayList<Statement>> statements = new HashMap<>();
-
 	long revisionId = 0;
 
+	
+	/**
+	 * Starts constructing an EntityDocument for the given entity id.
+	 * 
+	 * @param entityIdValue
+	 * 			the id of the document to build
+	 */
 	protected EntityDocumentBuilder(EntityIdValue entityIdValue) {
 		this.entityIdValue = entityIdValue;
 	}
@@ -70,23 +50,8 @@ public abstract class EntityDocumentBuilder<T extends EntityDocumentBuilder<T, O
 	protected EntityDocumentBuilder(O initialDocument) {
 		this.entityIdValue = initialDocument.getEntityId();
 		this.revisionId = initialDocument.getRevisionId();
-		for(MonolingualTextValue label : initialDocument.getLabels().values()) {
-			withLabel(label);
-		}
-		for(MonolingualTextValue description : initialDocument.getDescriptions().values()) {
-			withDescription(description);
-		}
-		for(List<MonolingualTextValue> aliases : initialDocument.getAliases().values()) {
-			for(MonolingualTextValue alias : aliases) {
-				withAlias(alias);
-			}
-		}
-		Iterator<Statement> iterator = initialDocument.getAllStatements();
-		while(iterator.hasNext()) {
-			withStatement(iterator.next());		
-		}
 	}
-
+	
 	/**
 	 * Sets the revision id for the constructed document. See
 	 * {@link EntityDocument#getRevisionId()}.
@@ -99,118 +64,4 @@ public abstract class EntityDocumentBuilder<T extends EntityDocumentBuilder<T, O
 		this.revisionId = revisionId;
 		return getThis();
 	}
-
-	/**
-	 * Adds an additional label to the constructed document.
-	 *
-	 * @param mtv
-	 *            the additional label
-	 * @return builder object to continue construction
-	 */
-	public T withLabel(MonolingualTextValue mtv) {
-		this.labels.add(mtv);
-		return getThis();
-	}
-
-	/**
-	 * Adds an additional label to the constructed document.
-	 *
-	 * @param text
-	 *            the text of the label
-	 * @param languageCode
-	 *            the language code of the label
-	 * @return builder object to continue construction
-	 */
-	public T withLabel(String text, String languageCode) {
-		withLabel(factory.getMonolingualTextValue(text, languageCode));
-		return getThis();
-	}
-
-	/**
-	 * Adds an additional description to the constructed document.
-	 *
-	 * @param mtv
-	 *            the additional description
-	 * @return builder object to continue construction
-	 */
-	public T withDescription(MonolingualTextValue mtv) {
-		this.descriptions.add(mtv);
-		return getThis();
-	}
-
-	/**
-	 * Adds an additional description to the constructed document.
-	 *
-	 * @param text
-	 *            the text of the description
-	 * @param languageCode
-	 *            the language code of the description
-	 * @return builder object to continue construction
-	 */
-	public T withDescription(String text, String languageCode) {
-		withDescription(factory.getMonolingualTextValue(text, languageCode));
-		return getThis();
-	}
-
-	/**
-	 * Adds an additional alias to the constructed document.
-	 *
-	 * @param mtv
-	 *            the additional alias
-	 * @return builder object to continue construction
-	 */
-	public T withAlias(MonolingualTextValue mtv) {
-		this.aliases.add(mtv);
-		return getThis();
-	}
-
-	/**
-	 * Adds an additional alias to the constructed document.
-	 *
-	 * @param text
-	 *            the text of the alias
-	 * @param languageCode
-	 *            the language code of the alias
-	 * @return builder object to continue construction
-	 */
-	public T withAlias(String text, String languageCode) {
-		withAlias(factory.getMonolingualTextValue(text, languageCode));
-		return getThis();
-	}
-
-	/**
-	 * Adds an additional statement to the constructed document.
-	 *
-	 * @param statement
-	 *            the additional statement
-	 * @return builder object to continue construction
-	 */
-	public T withStatement(Statement statement) {
-		PropertyIdValue pid = statement.getClaim().getMainSnak()
-				.getPropertyId();
-		ArrayList<Statement> pidStatements = this.statements.get(pid);
-		if (pidStatements == null) {
-			pidStatements = new ArrayList<Statement>();
-			this.statements.put(pid, pidStatements);
-		}
-
-		pidStatements.add(statement);
-		return getThis();
-	}
-
-	/**
-	 * Returns a list of {@link StatementGroup} objects for the currently stored
-	 * statements.
-	 *
-	 * @return
-	 */
-	protected List<StatementGroup> getStatementGroups() {
-		ArrayList<StatementGroup> result = new ArrayList<>(
-				this.statements.size());
-		for (ArrayList<Statement> statementList : this.statements.values()) {
-			result.add(factory.getStatementGroup(statementList));
-		}
-		return result;
-	}
-
 }
